+++ Test/Unit/Core/Api/EasyCreditHttpClientTest.php
