+++ Test/Unit/Core/Api/EasyCreditWebServiceClientTest.php
