+++ Test/Unit/Core/CrossCutting/EasyCreditLoggingTest.php
