+++ Test/Unit/Core/Di/EasyCreditApiConfigTest.php
