+++ Test/Unit/Core/Di/EasyCreditDicConfigTest.php
