+++ Test/Unit/Core/Domain/EasyCreditPaymentTest.php
