+++ Test/Unit/Core/Domain/EasyCreditSessionTest.php
