+++ Test/Unit/Core/Helper/EasyCreditHelperTest.php
