+++ Test/Unit/test-bin/03-texte_repraesentativesBeispiel.php
