+++ Test/Unit/test-bin/06-aquisition_value.php
