+++ Test/phpunit.xml
@@ -1,0 +1,59 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<phpunit
+    backupStaticAttributes="false"
+    cacheTokens="false"
+    colors="false"
+    convertErrorsToExceptions="true"
+    convertNoticesToExceptions="true"
+    convertWarningsToExceptions="true"
+    forceCoversAnnotation="false"
+    processIsolation="false"
+    stopOnError="false"
+    stopOnFailure="false"
+    stopOnIncomplete="false"
+    stopOnSkipped="false"
+    verbose="true"
+    backupGlobals="true"
+    bootstrap="bootstrap.php"
+    forceCoversAnnotation="false"
+    mapTestClassNameToCoveredClassName="false"
+    printerClass="PHPUnit\TextUI\ResultPrinter"
+    beStrictAboutTestsThatDoNotTestAnything="true"
+    strict="false"
+    >
+    <php>
+        <includePath>./tests</includePath>
+        <ini name="memory_limit" value="1024M"/>
+    </php>
+    <testsuites>
+        <testsuite name="AllTests">
+            <directory>./unit</directory>
+        </testsuite>
+    </testsuites>
+    <filter>
+        <whitelist addUncoveredFilesFromWhitelist="true">
+            <directory suffix=".php">../</directory>
+            <exclude>
+                <directory suffix=".php">../Tests/</directory>
+                <directory suffix=".php">../translations/</directory>
+                <directory suffix=".php">../views/</directory>
+                <directory suffix=".php">../out/</directory>
+                <directory suffix=".php">../docs/</directory>
+                <file suffix=".php">../metadata.php</file>
+            </exclude>
+        </whitelist>
+
+
+
+        <whitelist processUncoveredFilesFromWhitelist="true">
+            <directory suffix=".php">../</directory>
+            <exclude>
+                <directory suffix=".php">../unit</directory>
+            </exclude>
+        </whitelist>
+    </filter>
+    <logging>
+        <!--<log type="coverage-html" target="/tmp/report" charset="UTF-8"-->
+        <log type="coverage-clover" target="/tmp/report/clover.xml" lowUpperBound="35" highLowerBound="70"/>
+    </logging>
+</phpunit>
