--- tests/.gitignore
@@ -1,2 +1,0 @@
-/coverage/
-/easycredit.log
