--- tests/phpunit.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<phpunit
-        backupStaticAttributes="false"
-        cacheTokens="false"
-        colors="false"
-        convertErrorsToExceptions="true"
-        convertNoticesToExceptions="true"
-        convertWarningsToExceptions="true"
-        forceCoversAnnotation="false"
-        processIsolation="false"
-        stopOnError="false"
-        stopOnFailure="false"
-        stopOnIncomplete="false"
-        stopOnSkipped="false"
-        verbose="true"
-        >
-    <php>
-        <includePath>./tests</includePath>
-        <ini name="memory_limit" value="1024M"/>
-    </php>
-    <testsuites>
-        <testsuite name="AllTests">
-            <directory>./unit</directory>
-        </testsuite>
-    </testsuites>
-    <filter>
-        <whitelist processUncoveredFilesFromWhitelist="true">
-            <directory suffix=".php">../../</directory>
-            <exclude>
-                <directory suffix=".php">./unit</directory>
-            </exclude>
-        </whitelist>
-    </filter>
-    <logging>
-        <!--<log type="coverage-html" target="/tmp/report" charset="UTF-8"-->
-        <log type="coverage-clover" target="/tmp/report/clover.xml" lowUpperBound="35" highLowerBound="70"/>
-    </logging>
-</phpunit>
